backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.29.0
pandas==2.2.1
scikit-learn==1.4.1.post1
numpy==1.26.4
python-multipart==0.0.9
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pydantic[email]==2.6.4
